Add alias python=python3
If finds python3, but can't find python, make an alias. I mean, ideally
that would been a package on all available systems like
python-is-python3 on Ubuntu, or just give me python as python3, python2
was deprecated like ten years ago if not more.

But yeah, alias it is.

I also added some pretty loading messages when first starting zsh and it
attempts to load grml configuration and agkozak prompt.